Move recommendations UI to recommendations section
