+++ lib/data/bloc/recommendations/RecommendationCardTransformer.dart
@@ -1,12 +1,10 @@
 import 'package:farmsmart_flutter/data/bloc/recommendations/RecommendationEngine.dart';
 import 'package:farmsmart_flutter/data/model/crop_entity.dart';
-import 'package:farmsmart_flutter/ui/common/recommendation_card/recommendation_card_view_model.dart';
-import 'package:farmsmart_flutter/ui/crop/viewmodel/CropDetailViewModel.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_card/recommendation_card_view_model.dart';
 import 'package:intl/intl.dart';
 
 import '../Basket.dart';
 import '../Transformer.dart';
-import '../ViewModelProvider.dart';
 
 class _Constants {
   static final cent = 100.0;
