+++ lib/ui/crop/CropInfoList.dart
@@ -1,6 +1,6 @@
 import 'package:farmsmart_flutter/ui/common/SectionListView.dart';
 import 'package:farmsmart_flutter/ui/common/headerAndFooterListView.dart';
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_listitem/recommendation_detail_listitem.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_listitem/recommendation_detail_listitem.dart';
 import 'package:flutter/widgets.dart';
 
 class CropInfoList extends StatelessWidget implements  ListViewSection {
