+++ lib/ui/playground/data/playground_recommendation_detail_card_datasource.dart
@@ -1,8 +1,8 @@
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_card/mock_recommendation_detail_card_view_model.dart';
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_card/recommendation_detail_card.dart';
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_card/recommendation_detail_card_styles.dart';
 import 'package:farmsmart_flutter/ui/playground/data/playground_data_source.dart';
 import 'package:farmsmart_flutter/ui/playground/playground_widget.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_card/mock_recommendation_detail_card_view_model.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_card/recommendation_detail_card.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_card/recommendation_detail_card_styles.dart';
 import 'package:flutter/widgets.dart';
 
 class PlaygroundRecommendationDetailCardDatasource
