+++ lib/ui/recommendations/recommendation_card/recommendation_card.dart
@@ -1,9 +1,10 @@
-import 'package:farmsmart_flutter/ui/common/recommendation_card/recommendation_card_view_model.dart';
 import 'package:farmsmart_flutter/ui/common/roundedButton.dart';
 import 'package:farmsmart_flutter/ui/common/rounded_button_stateful.dart';
 import 'package:farmsmart_flutter/ui/common/rounded_image_overlay.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+
+import 'recommendation_card_view_model.dart';
 
 export 'package:farmsmart_flutter/ui/common/roundedButton.dart';
