+++ lib/ui/recommendations/recommendation_card/recommendation_card_view_model.dart
