+++ lib/ui/recommendations/recommendation_compact_card/recommendation_compact_card.dart
@@ -1,7 +1,7 @@
-import 'package:farmsmart_flutter/ui/common/recommendation_card/recommendation_card_view_model.dart';
 import 'package:farmsmart_flutter/ui/common/roundedButton.dart';
 import 'package:farmsmart_flutter/ui/common/rounded_button_stateful.dart';
 import 'package:farmsmart_flutter/ui/common/rounded_image_overlay.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_card/recommendation_card_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
