+++ lib/ui/recommendations/recommendation_detail_card/mock_recommendation_detail_card_view_model.dart
