+++ lib/ui/recommendations/recommendation_detail_listitem/mock_recommendation_detail_listitem_view_model.dart
