+++ test/ui/recommendation/recommendation_card/recommendation_card_add_to_plot_test.png
