+++ test/ui/recommendation/recommendation_card/recommendation_card_added_to_plot_test.png
