+++ test/ui/recommendation/recommendation_compact_card/recommendation_compact_card_add_to_plot_test.png
