+++ test/ui/recommendation/recommendation_compact_card/recommendation_compact_card_added_to_plot_test.png
