+++ test/ui/recommendation/recommendation_detail_card/recommendation_detail_card_add_to_plot_test.png
