+++ test/ui/recommendation/recommendation_detail_listitem/recommendation_detail_listitem_test.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_listitem/mock_recommendation_detail_listitem_view_model.dart';
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_listitem/recommendation_detail_listitem.dart';
-import 'package:farmsmart_flutter/ui/common/recommendation_detail_listitem/recommendation_detail_listitem_styles.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_listitem/mock_recommendation_detail_listitem_view_model.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_listitem/recommendation_detail_listitem.dart';
+import 'package:farmsmart_flutter/ui/recommendations/recommendation_detail_listitem/recommendation_detail_listitem_styles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
