+++ test/ui/recommendation/recommendation_detail_listitem/recommendation_detail_listitem_with_color_list.png
